Fixed a bug that caused strange colors when 8bitColor was specified.

+++ src/Effect.h
@@ -14,24 +14,24 @@ namespace m5avatar {
 class Effect final : public Drawable {
  private:
   void drawBubbleMark(M5Canvas *spi, uint32_t x, uint32_t y, uint32_t r,
-                      uint32_t color) {
+                      uint16_t color) {
     drawBubbleMark(spi, x, y, r, color, 0);
   }
 
   void drawBubbleMark(M5Canvas *spi, uint32_t x, uint32_t y, uint32_t r,
-                      uint32_t color, float offset) {
+                      uint16_t color, float offset) {
     r = r + floor(r * 0.2 * offset);
     spi->drawCircle(x, y, r, color);
     spi->drawCircle(x - (r / 4), y - (r / 4), r / 4, color);
   }
 
   void drawSweatMark(M5Canvas *spi, uint32_t x, uint32_t y, uint32_t r,
-                 uint32_t color) {
+                 uint16_t color) {
     drawSweatMark(spi, x, y, r, color, 0);
   }
 
   void drawSweatMark(M5Canvas *spi, uint32_t x, uint32_t y, uint32_t r,
-                 uint32_t color, float offset) {
+                 uint16_t color, float offset) {
     y = y + floor(5 * offset);
     r = r + floor(r * 0.2 * offset);
     spi->fillCircle(x, y, r, color);
@@ -41,12 +41,12 @@ class Effect final : public Drawable {
   }
 
   void drawChillMark(M5Canvas *spi, uint32_t x, uint32_t y, uint32_t r,
-                     uint32_t color) {
+                     uint16_t color) {
     drawChillMark(spi, x, y, r, color, 0);
   }
 
   void drawChillMark(M5Canvas *spi, uint32_t x, uint32_t y, uint32_t r,
-                     uint32_t color, float offset) {
+                     uint16_t color, float offset) {
     uint32_t h = r + abs(r * 0.2 * offset);
     spi->fillRect(x - (r / 2), y, 3, h / 2, color);
     spi->fillRect(x, y, 3, h * 3 / 4, color);
@@ -54,12 +54,12 @@ class Effect final : public Drawable {
   }
 
   void drawAngerMark(M5Canvas *spi, uint32_t x, uint32_t y, uint32_t r,
-                     uint32_t color, uint32_t bColor) {
+                     uint16_t color, uint32_t bColor) {
     drawAngerMark(spi, x, y, r, color, bColor, 0);
   }
 
   void drawAngerMark(M5Canvas *spi, uint32_t x, uint32_t y, uint32_t r,
-                     uint32_t color, uint32_t bColor, float offset) {
+                     uint16_t color, uint32_t bColor, float offset) {
     r = r + abs(r * 0.4 * offset);
     spi->fillRect(x - (r / 3), y - r, (r * 2) / 3, r * 2, color);
     spi->fillRect(x - r, y - (r / 3), r * 2, (r * 2) / 3, color);
@@ -68,12 +68,12 @@ class Effect final : public Drawable {
   }
 
   void drawHeartMark(M5Canvas *spi, uint32_t x, uint32_t y, uint32_t r,
-                 uint32_t color) {
+                 uint16_t color) {
     drawHeartMark(spi, x, y, r, color, 0);
   }
 
   void drawHeartMark(M5Canvas *spi, uint32_t x, uint32_t y, uint32_t r,
-                 uint32_t color, float offset) {
+                 uint16_t color, float offset) {
     r = r + floor(r * 0.4 * offset);
     spi->fillCircle(x - r / 2, y, r / 2, color);
     spi->fillCircle(x + r / 2, y, r / 2, color);
@@ -90,8 +90,8 @@ class Effect final : public Drawable {
   Effect(const Effect &other) = default;
   Effect &operator=(const Effect &other) = default;
   void draw(M5Canvas *spi, BoundingRect rect, DrawContext *ctx) override {
-    uint32_t primaryColor = ctx->getColorDepth() == 1 ? 1 : ctx->getColorPalette()->get(COLOR_PRIMARY);
-    uint32_t bgColor = ctx->getColorDepth() == 1 ? ERACER_COLOR : ctx->getColorPalette()->get(COLOR_BACKGROUND);
+    uint16_t primaryColor = ctx->getColorDepth() == 1 ? 1 : ctx->getColorPalette()->get(COLOR_PRIMARY);
+    uint16_t bgColor = ctx->getColorDepth() == 1 ? ERACER_COLOR : ctx->getColorPalette()->get(COLOR_BACKGROUND);
     float offset = ctx->getBreath();
     Expression exp = ctx->getExpression();
     switch (exp) {
